requirements.txt
groq
python-dotenv
sounddevice
scipy
numpy
google-cloud-texttospeech
pygame
openai